[TM-1781] Simplify the setup of REPL for new services.
