+++ apps/unified-database-service/src/repl.ts
@@ -1,20 +1,4 @@
-import { existsSync, mkdirSync } from "node:fs";
-import { join } from "node:path";
-import { repl } from "@nestjs/core";
 import { AppModule } from "./app.module";
-import { TMLogger } from "@terramatch-microservices/common/util/tm-logger";
+import { bootstrapRepl } from "@terramatch-microservices/common/util/bootstrap-repl";
 
-const logger = new TMLogger("REPL");
-
-async function bootstrap() {
-  const replServer = await repl(AppModule);
-
-  const cacheDirectory = join("node_modules", ".cache");
-
-  if (!existsSync(cacheDirectory)) mkdirSync(cacheDirectory);
-
-  replServer.setupHistory(join(cacheDirectory, ".nestjs_repl_history"), error => {
-    if (error) logger.error(error);
-  });
-}
-bootstrap();
+bootstrapRepl("Unified Database Service", AppModule);
