+++ apps/user-service/project.json
@@ -7,13 +7,9 @@
   "targets": {
     "build-repl": {
       "executor": "@nx/webpack:webpack",
-      "outputs": ["{options.outputPath}"],
       "options": {
-        "webpackConfig": "apps/user-service/repl.webpack.config.js",
-        "target": "node",
-        "compiler": "tsc",
-        "outputPath": "dist/apps/user-service-repl",
-        "main": "apps/user-service/src/repl.ts"
+        "webpackConfig": "repl/webpack.config.js",
+        "outputPath": "{workspaceRoot}/dist/{projectRoot}-repl"
       }
     },
     "serve": {
